feat: moves generation action content from modal to drawer component
