+++ app/src/i18n/locales/fr/hiap.json
@@ -47,7 +47,7 @@
   "top-actions-for-your-city-description": "Voici les 40 principales actions que nous recommandons basées sur votre inventaire GES et les données de contexte de votre ville. Vous trouverez 20 actions pour l'Atténuation et 20 pour l'Adaptation dans leurs onglets respectifs. Il est recommandé de relancer le priorisateur après toute mise à jour ou modification pertinente des données.",
   "actions-for-your-city-description": "Voici les 40 principales actions que nous recommandons basées sur votre inventaire GES, l'Évaluation des Risques Climatiques et les données de contexte de votre ville. Vous trouverez 20 actions pour l'Atténuation et 20 pour l'Adaptation dans leurs onglets respectifs.",
   "re-prioritize-actions": "Reprioriser les actions",
-  "generated-action-plan": "générer plan d'action",
+  "generated-action-plan": "Générer Plan d'Action",
   "export-as-pdf": "exporter en PDF",
   "generating-pdf": "génération du PDF...",
   "subactions": "Sous-actions",
